anki-python/requirements.txt
Markdown==2.6.11
pyyaml==5.3.1
